requirements.txt
google-cloud-secret-manager==2.16.0
google-cloud-storage==2.13.0
requests==2.31.0
python-dotenv==1.0.1
fit-tool==0.9.15
garminconnect==0.2.14